Add files via upload
This is based on a follow along youtube project that I did. I am really proud of how this turned out and some of the skills that I have learned and used: XLOOKUP, INDEX MATCH, as well as honing my pivot table skills, and using this as well as slicers to create a dashboard.