Add files via upload
This was also following along with the video, whilst also creating some of my own ideas for the data analysis and visualisation. The cleaning involved some data verification, as well as standardising some of the data and using search special to change the name of a column from teamgame to team, as well as some other bits. From there I did some data analysis and created some pivot tables as well as using slicers.