Add files via upload
This is some work I did with python. I tried my best to work through it on my own, but I used the 'How to Build a Robo-Advisor With Python' book to help create an understanding as well as to consult when encountering problems. This combined my knowledge of modern portfolio theory with my learning in python.